+++ Assets/MoreHit/Scripts/InGame/Enemy/Types/ZakoEnemy.cs
@@ -35,7 +35,7 @@ namespace MoreHit.Enemy
         /// <summary>
         /// ダメージを受けた時の雑魚敵固有の処理
         /// </summary>
-        protected override void OnDamageReceived(float damage)
+        protected override void OnDamageReceived(int damage)
         {
             base.OnDamageReceived(damage);
             // TODO: ダメージ時のエフェクト、アニメーション、ノックバックなどを実装
